requirements.txt
alembic==1.12.0
appdirs==1.4.4
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.1.1
cycler==0.10.0
databricks-cli==0.18.0
defusedxml==0.7.1
distlib==0.3.7
docker==6.1.3
entrypoints==0.4
exceptiongroup==1.1.2
filelock==3.12.4
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
fonttools==4.43.1
frozendict==2.3.8
fsspec==2023.9.2
gitdb==4.0.10
GitPython==3.1.40
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
idna==2.10
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lxml==4.9.3
Mako==1.2.4
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.8.0
mlflow==2.7.1
mpmath==1.3.0
multitasking==0.0.11
names==0.3.0
networkx==3.1
numpy==1.25.2
oauthlib==3.2.2
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.74
outcome==1.2.0
packaging==23.2
pandas==2.1.1
peewee==3.16.3
Pillow==10.1.0
platformdirs==3.11.0
protobuf==4.24.4
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==13.0.0
pygame==2.5.2
PyJWT==2.8.0
pyparsing==3.1.1
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
querystring-parser==1.2.4
requests==2.31.0
requests-toolbelt==1.0.0
roboflow==1.1.19
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.13.2
selenium==4.10.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.17
sqlparse==0.4.4
supervision==0.18.0
sympy==1.12
tabulate==0.9.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tk==0.1.0
torch==2.2.0
torchsummary==1.5.1
torchvision==0.17.0
tqdm==4.66.2
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.9.0
tzdata==2023.3
ultralytics==8.0.196
urllib3==2.0.3
virtualenv==20.24.5
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==2.3.6
wsproto==1.2.0
yfinance==0.2.30
zipp==3.17.0
